pyproject.toml
[project]
name = "mining-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.5.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
]
